UPDATE: Create helpers for my overtime logic
